apps/api/requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
onnxruntime==1.19.*
numpy==2.*
opencv-python-headless==4.10.*
pillow==10.*
psycopg2-binary==2.9.*
python-multipart==0.0.9
